Add 'split_on_note' config option to migration script
This option when enabled will always create a sublist upon finding a
'public note' field within a list and assign items following it to that
sublist instead of the parent list.

This is for the case where users have been working around the lack of
sublists in rl1 by using notes as category headings.